Implement router how we intended and handle login states
