+++ src/Action-creators/toggleFavorite.js
@@ -1,5 +1,4 @@
 export const toggleFavorite = id => {
-  console.log(id)
   return {
   type: 'TOGGLE_FAVORITE',
   id,
