+++ src/containers/App/App.js
@@ -5,17 +5,16 @@ import CardContainer from '../CardContainer/CardContainer'
 import './App.css';
 import SignUp from '../Authentication/SignUp/SignUp';
 import SignIn from '../Authentication/SignIn/SignIn';
+import { logOut } from '../../Action-creators/logOut';
 
 class App extends Component {
 
   render() {
-    const { id, name } = this.props;
+    const { id, name, logoutUser } = this.props;
     return (
       <div className="App">
-        {id ? <h2>{`Welcome back, ${name}!`}</h2> : <h2>Please sign in</h2>}
-        <NavLink to='/login'>
-          Login 
-        </NavLink>
+        {id ? <h2>{`Welcome, ${name}!`}</h2> : <h2>Please sign in</h2>}
+        {!id ? <NavLink to='/login'>Login</NavLink> : <NavLink to='/' onClick={logoutUser}>Sign Out</NavLink>}
         <Route exact path='/login' component={ SignIn } />
         <Route exact path='/signup' component={ SignUp } />
         <Route path='/' component={ CardContainer }/>
@@ -24,9 +23,13 @@ class App extends Component {
   }
 }
 
-const mapStateToProps = (state) => ({
+export const mapStateToProps = (state) => ({
   id: state.id,
   name: state.name,
 })
 
-export default withRouter(connect(mapStateToProps, null )(App));
+export const mapDispatchToProps = dispatch => ({
+  logoutUser: () => dispatch(logOut())
+})
+
+export default withRouter(connect(mapStateToProps, mapDispatchToProps)(App));
